Implemented user location access

+++ app/src/main/java/com/example/chaos/Data/Wind.java
@@ -2,13 +2,13 @@ package com.example.chaos.Data;
 
 public class Wind{
 	private int deg;
-	private int speed;
+	private double speed;
 
 	public int getDeg(){
 		return deg;
 	}
 
-	public int getSpeed(){
+	public double getSpeed(){
 		return speed;
 	}
 }
